printf

+++ ft_putptr.c
@@ -6,13 +6,13 @@
 /*   By: souel-bo <souel-bo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/24 11:57:15 by souel-bo          #+#    #+#             */
-/*   Updated: 2024/11/24 13:10:08 by souel-bo         ###   ########.fr       */
+/*   Updated: 2024/11/24 13:55:46 by souel-bo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-void	ft_putptr(unsigned long long nb, int *counter)
+void	ft_putptr(unsigned long nb, int *counter)
 {
 	if (nb == 0)
 	{
